Move mock vws
